refactor(prices): move currency toggle into the header row
Lift the "In {base} / Original quote" segment control out of the table
view and into the page header, right-aligned next to the title, shown
only on the Known prices tab. Extracted as PriceCurrencyToggle.
